EID-1117 - Turn DuplicateAssertionValidator into interface
- So it can be used with a different implementation in proxy node
